backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
groq>=0.4.0
numpy>=1.24.0

python-dotenv>=1.0.0
gTTS>=2.3.0
pydub>=0.25.0
